backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
PyJWT>=2.8.0
mangum>=0.17.0
